Add files via upload
Write a bubble sort program in Python 3. A bubble sort is an algorithm that will sort a list of values into order. On each pass the bubble sort passes through a list of values from start to finish comparing each sequential pair of values. If the values compared are not in order, they are swapped. The sort begins at the start of the list, first comparing the values in position 0 and 1, next the values in positions 1 and 2, then those in positions 2 and 3, and so on until the end of the list is reached. At the end of each pass if the entire list is not in the correct order, the sort will start at the beginning of the list and pass through it again. The sort will continue its passes through the list until the list is sorted.

The program should prompt the user to input a list of comma-separated integers and should sort the list into ascending order. For instance, if the user inputs 4, 9, 74, 0, 9, 8, 28, 1 this should be sorted to 0, 1, 4, 8, 9, 9, 28, 74

If the user inputs the values in the wrong format the program should detect this and prompt the user to input data in the correct format.

Your program should display the original list on the first line and then each time the sort reaches the end of the list the program should display the number of the pass and the list as it is currently sorted.

Once the sort is complete, the original unsorted list and the sorted list should be displayed along with the number of passes it required to completely sort the list.